prune: remove unused code
