+++ packages/android-performance-profiler/src/commands/getAbi.ts
@@ -1,4 +1,4 @@
-import { executeCommand } from "./shellNext";
+import { executeCommand } from "./shell";
 
 export const getAbi = () =>
   executeCommand("adb shell getprop ro.product.cpu.abi").split("\n")[0];
